requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
blinker==1.5
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
ez-setup==0.9
Flask==2.2.3
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-WTF==1.0.1
FlightRadarAPI==1.2.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
load-dotenv==0.1.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.1
nltk==3.8.1
-e git+https://github.com/openskynetwork/opensky-api.git@0a4259acee4e17d7e9b2f88e1d3c59d52f01afbf#egg=opensky_api&subdirectory=python
PyMySQL==1.0.2
python-dotenv==0.21.0
regex==2022.10.31
requests==2.28.1
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
skyscanner==1.1.5
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.15.0
